+++ response.go
@@ -141,7 +141,10 @@ func (r *Response) Download(file string, fileName string) error {
 // Error print the error in the console and return it with an error code 500.
 // If debugging is enabled in the config, the error is also written in the response
 // and the stacktrace is printed in the console.
+// If debugging is not enabled, only the stauts code is set, which means you can still
+// write to the response, or use your error status handler.
 func (r *Response) Error(err interface{}) error {
+	r.err = err
 	dbg := config.GetBool("debug")
 	log.Println(err)
 	if dbg {
@@ -155,8 +158,8 @@ func (r *Response) Error(err interface{}) error {
 		return r.JSON(http.StatusInternalServerError, map[string]interface{}{"error": message})
 	}
 
-	r.empty = false
-	r.WriteHeader(http.StatusInternalServerError)
+	// Don't set r.empty to false to let error status handler process the error
+	r.Status(http.StatusInternalServerError)
 	return nil
 }